Single Tenant, Multi-Company  (#124)
* feat: migration to support multiple companies

* feat: add companyId to permissions model

* refactor: create a company RLS

* refactor: migrations for multi-tenant - batch 1

* refactor: company form to support mult-tenant

* refactor: update requirePermissions check to include current company

* feat: add edge function to seed company when its created

* fix: functions to evaluate company claims for RLS

* feat: add ability to run redis locally

* refactor: migrations for multi-tenant - batch 2

* refactor: migrations for multi-tenant - batch 3

* refactor: migrations for multi-tenant - batch 4

* refactor: migrations for multi-tenant - batch 5

* refactor: migrations for multi-tenant - batch 7

* refactor: migrations for multi-tenant - batch 8

* refactor: migrations for multi-tenant - batch 9

* refactor: migrations for multi-tenant - batch 10

* refactor: migrations for multi-tenant - batch 11

* refactor: migrations for multi-tenant - batch 12

* refactor: migrations for multi-tenant - batch 13

* refactor: migrations for multi-tenant - batch 14

* refactor: migrations for multi-tenant - batch 15

* refactor: migrations for multi-tenant - batch 16

* build: hard-code meta tags in root

* refactor: services for multi-tenant - resources

* refactor: services for multi-tenant - parts

* refactor: services for multi-tenant - accounting

* refactor: move employee type and permissions generation from seed.ts to seed-company function

* refactor: services for multi-tenant - inventory

* refactor: services for multi-tenant - purchasing

* refactor: services for multi-tenant - sales

* refactor: services for multi-tenant - inventory and documents

* refactor: services for multi-tenant - settings

* refactor: services for multi-tenant - users

* refactor: cleanup services

* refactor: triggers for multi-tenant

* fix: employee types and user management

* refactor: remove cookie files

* fix: account create for users, customers, and suppliers

* refactor: fix triggers for multi-tenant

* refactor: frontend supabase client queries and mutations

* reafactor: companyId foregin key function

* reafactor: add companyId to edge functions

* feat: ui for company dropdown in breadcrumbs

* feat: new company form, and separating permissions into users table

* refactor: move permissions to user table and out of jwt

* feat: company switching

* docs: update docs for multi-company permissions

* fix: problems with parts module

* Company id to string (#136)

* 1/3

* more updates

* refactor: use xid for company id instead of number

---------


* fix: don't create duplicate userToCompany entries

* docs: update to elastic license

* fix: bug fixes in invite/company creation

* docs: add an update to the website about the elastic license

* seed currencies and custom fields in migration file

* fix: custom fields being company scoped, but tables being global

* fix: integrations and currencies for multi-company

* fix: sales orders leaking across companies

* style: fix scroll issue on drawer

* refactor: move permissions into its own table

* feat: add query to fetch invitable users

* feat: add ui for inviting existing users from other companies

* fix: employee jobs

* fix: customers and suppliers can CRUD documents

* fix: date format

* fix: default accounts

* fix: accounting problems

* fix: user groups leaking

* fix: minor updates

---------

+++ apps/carbon/app/modules/documents/ui/Documents/useDocument.ts
@@ -41,6 +41,8 @@ export const useDocument = () => {
   const insertTransaction = useCallback(
     (document: DocumentType, type: DocumentTransactionType) => {
       if (user?.id === undefined) throw new Error("User is undefined");
+      if (!document.id) throw new Error("Document id is undefined");
+
       return supabase?.from("documentTransaction").insert({
         documentId: document.id,
         type,
@@ -52,11 +54,14 @@ export const useDocument = () => {
 
   const deleteLabel = useCallback(
     async (document: DocumentType, label: string) => {
+      if (!document.id) throw new Error("Document id is undefined");
+      if (user?.id === undefined) throw new Error("User is undefined");
+
       return supabase
         ?.from("documentLabel")
         .delete()
         .eq("documentId", document.id)
-        .eq("userId", user?.id)
+        .eq("userId", user.id)
         .eq("label", label);
     },
     [supabase, user?.id]
@@ -64,6 +69,8 @@ export const useDocument = () => {
 
   const download = useCallback(
     async (doc: DocumentType) => {
+      if (!doc.path) throw new Error("Document path is undefined");
+
       const result = await supabase?.storage.from("private").download(doc.path);
 
       if (!result || result.error) {
@@ -75,7 +82,7 @@ export const useDocument = () => {
       document.body.appendChild(a);
       const url = window.URL.createObjectURL(result.data);
       a.href = url;
-      a.download = doc.name;
+      a.download = doc.name ?? "File";
       a.click();
 
       setTimeout(() => {
@@ -97,7 +104,7 @@ export const useDocument = () => {
 
   const edit = useCallback(
     (document: DocumentType) =>
-      navigate(`${path.to.document(document.id)}?${params}`),
+      navigate(`${path.to.document(document.id!)}?${params}`),
     [navigate, params]
   );
 
@@ -107,13 +114,13 @@ export const useDocument = () => {
         await supabase
           ?.from("documentFavorite")
           .delete()
-          .eq("documentId", document.id)
+          .eq("documentId", document.id!)
           .eq("userId", user?.id);
         return insertTransaction(document, "Unfavorite");
       } else {
         await supabase
           ?.from("documentFavorite")
-          .insert({ documentId: document.id, userId: user?.id });
+          .insert({ documentId: document.id!, userId: user?.id });
         return insertTransaction(document, "Favorite");
       }
     },
@@ -136,12 +143,12 @@ export const useDocument = () => {
       await supabase
         ?.from("documentLabel")
         .delete()
-        .eq("documentId", document.id)
+        .eq("documentId", document.id!)
         .eq("userId", user.id)
         .then(() => {
           return supabase?.from("documentLabel").insert(
             labels.map((label) => ({
-              documentId: document.id,
+              documentId: document.id!,
               label,
               userId: user.id,
             }))
@@ -155,6 +162,7 @@ export const useDocument = () => {
 
   const makePreview = useCallback(
     async (doc: DocumentType) => {
+      if (!doc.path) throw new Error("Document path is undefined");
       const result = await supabase?.storage.from("private").download(doc.path);
 
       if (!result || result.error) {
@@ -172,7 +180,7 @@ export const useDocument = () => {
       return supabase
         ?.from("documentLabel")
         .delete()
-        .eq("documentId", document.id)
+        .eq("documentId", document.id!)
         .eq("userId", user?.id)
         .eq("label", label);
     },
